leaderboard delete button fixed and overwrite implemented

+++ ProjectVizuelno/Start.cs
@@ -23,17 +23,16 @@ namespace ProjectVizuelno
         {
             InitializeComponent();
             this.CenterToScreen();
+            this.FormBorderStyle = FormBorderStyle.FixedSingle;
+            this.MaximizeBox = false;
             GC.Collect();
             Form1 game1 = new Form1(0,"");
             Form2 game2 = new Form2(0,"");
-            /*this.FormBorderStyle = FormBorderStyle.FixedSingle;
-            this.MaximizeBox = false;
-            this.MinimizeBox = false; */
-            /*if (!(File.Exists("leaderboard.txt")))  // ENABLE THIS WHEN EVERYTHING ELSE IS DONE
+            if (!(File.Exists("leaderboard.txt")))  // ENABLE THIS WHEN EVERYTHING ELSE IS DONE
             {
                 var newFile = File.Create("leaderboard.txt");
                 newFile.Close();
-            } */
+            } 
         }
 
         private bool tezina = false; // bool promenliva za dali e stisnato na edna od tezinite
@@ -74,10 +73,6 @@ namespace ProjectVizuelno
             else if (textBox1.TextLength <=2)
             {
                 MessageBox.Show("Вашето име е прекратко!", "Грешка", MessageBoxButtons.OK);
-            }
-            else if (textBox1.TextLength > 6)
-            {
-                MessageBox.Show("Вашето име е предолго. Максимум е 6 карактери!", "Грешка", MessageBoxButtons.OK);
             }
             else if(level==1 || level==2)
             {
